Relocate error messages into utility file
